% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class001-bids_class_base.R
\name{bids_class_base}
\alias{bids_class_base}
\title{Base abstract class}
\usage{
bids_class_base()
}
\description{
All \code{bidsr} classes should inherit this abstract class, which provides
consistent \code{S3} generics. This is an abstract class, which should not
be called directly; use \code{\link{new_bids_class}} to create new class
definitions.
}
