% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class02-bids_map.R
\name{bids_map}
\alias{bids_map}
\title{Low-level nested map to store key-value data with inherited structure}
\usage{
bids_map(parent = NULL, search_depth = BIDS_MAP_MAX_DEPTH())
}
\arguments{
\item{parent}{\code{NULL} if the map is at the top level, or another
map to inherit}

\item{search_depth}{integer maximum search depths; default is \code{29};
set options \code{'bidsr.map.search_depth'} or environment variable
\code{'BIDS_MAP_MAX_DEPTH'} to change the default depth}
}
\value{
A \code{'bids_map'} object.
}
\description{
Low-level nested map to store key-value data with inherited structure
}
\examples{


root_map <- bids_map()
root_map$key1 <- 1
root_map$key2 <- 2
names(root_map)

child_map <- bids_map(parent = root_map)
child_map$key3 <- 3
names(child_map)
child_map$key1
child_map$key2

# mask key2
child_map$key2 <- "a"
child_map

root_map$key2
child_map$key2

# nested maps
grand_child <- bids_map(parent = child_map)

# value comes from child map
grand_child$key2

# remove key2 from child map
child_map@impl$remove("key2")

# key2 is from root map now
grand_child$key2


}
