% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class031-bids_participants.R
\name{bids_tabular_participants}
\alias{bids_tabular_participants}
\title{'BIDS' participant table class}
\usage{
bids_tabular_participants(content, meta = NULL)
}
\arguments{
\item{content, meta}{see \code{\link{bids_tabular}}}
}
\value{
A \code{bids_tabular_participants} instance inheriting
\code{\link{bids_tabular}}.
}
\description{
A tabular containing a list of participants and their demographics.
The class is a child class of \code{\link{bids_tabular}}, hence see
the methods there.
The original specification is at
\url{https://bids-specification.readthedocs.io/en/stable/modality-agnostic-files.html#participants-file}.
}
\examples{



# basic
tabular <- bids_tabular_participants(
  data.frame(
    participant_id = "sub-001"
  )
)
tabular


# Run `download_bids_examples()` first
examples <- download_bids_examples(test = TRUE)
if(!isFALSE(examples)) {

  file <- file.path(examples, "ieeg_epilepsy_ecog", "participants.tsv")

  # read tabular as bids_tabular_participants
  as_bids_tabular(file, cls = bids_tabular_participants)

  # convert existing tabular
  tabular <- bids_tabular(
    data.frame(
      participant_id = "sub-001"
    )
  )
  tabular <- as_bids_tabular(tabular, cls = bids_tabular_participants)

  # save to tsv
  tsv <- file.path(tempdir(), "participants.tsv")
  paths <- tabular$save(tsv)
  print(paths)

  # use base R to read
  read.table(tsv, header = TRUE, na.strings = "n/a")

  # get sidecar
  cat(readLines(paths$sidecar_path), sep = "\n")

  unlink(tsv)
  unlink(paths$sidecar_path)
}

}
