% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class022-bids_samples.R
\name{bids_tabular_samples}
\alias{bids_tabular_samples}
\title{'BIDS' samples table class}
\usage{
bids_tabular_samples(content, meta = NULL)
}
\arguments{
\item{content, meta}{see \code{\link{bids_tabular}}}
}
\value{
A \code{bids_tabular_samples} instance inheriting
\code{\link{bids_tabular}}.
}
\description{
A tabular containing a list of samples and their metadata.
The class is a child class of \code{\link{bids_tabular}}, hence see
the methods there.
The original specification is at
\url{https://bids-specification.readthedocs.io/en/stable/modality-agnostic-files.html#samples-file}.
}
\examples{


# basic
tabular <- bids_tabular_samples(
  data.frame(
    sample_id = "sample-001",
    participant_id = "sub-001",
    sample_type = "cell line"
  )
)
tabular


# convert existing tabular
tabular <- bids_tabular(
  data.frame(
    sample_id = "sample-001",
    participant_id = "sub-001",
    sample_type = "cell line"
  )
)
tabular <- as_bids_tabular(tabular, cls = bids_tabular_samples)


# save to tsv
tsv <- file.path(tempdir(), "samples.tsv")
paths <- tabular$save(tsv)
print(paths)

# use base R to read
read.table(tsv, header = TRUE, na.strings = "n/a")

# get sidecar
cat(readLines(paths$sidecar_path), sep = "\n")

# clean up
unlink(tsv)
unlink(paths$sidecar_path)


}
