% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class005-bids_uri.R
\name{bids_uri}
\alias{bids_uri}
\title{'BIDS' uniform resource indicator ('URI') class definition}
\usage{
bids_uri(uri)
}
\arguments{
\item{uri}{'URI' string or another 'BIDS-URI' object}
}
\value{
A \code{bids_uri} instance.
}
\description{
'BIDS' uniform resource indicator ('URI') class definition
}
\examples{

# basic properties
uri <- bids_uri("bids::sub-01/fmap/sub-01_dir-AP_epi.nii.gz")
uri
uri$relative_path
uri$dataset_name

# set the entire uri
uri$format <- "bids:deriv1:sub-02/anat/sub-02_T1w.nii.gz"
uri

# resolve BIDS URI (partial support)


# resolving a BIDS URI requires dataset_description.json
data_description <- dataset_description <- as_bids_dataset_description(
  parent_directory = "/path/to/BIDS/folder",
  Name = "A dummy experiments",
  BIDSVersion = "1.6.0",

  DatasetLinks = list(
    "deriv1" = "derivatives/derivative1",
    "phantoms" = "file:///data/phantoms"
  )
)

uri <- bids_uri("bids::sub-01/fmap/sub-01_dir-AP_epi.nii.gz")
resolved <- uri$resolve(data_description)

# `raw_resolution` is relative to the parent directory where
# `dataset_description.json` is stored
resolved$raw_resolution
resolved$absolute_path

uri <- bids_uri("bids:deriv1:sub-02/anat/sub-02_T1w.nii.gz")
resolved <- uri$resolve(data_description)
resolved$raw_resolution
resolved$absolute_path


}
