% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class002-bids_class_base.R
\name{BIDSClassBase}
\alias{BIDSClassBase}
\title{Low-level abstract class for \code{bidsr}}
\usage{
BIDSClassBase()
}
\value{
Do not call this \code{S7} class directly, see
\code{\link{new_bids_class}} on how to use it properly
}
\description{
Low-level abstract class definition; see \code{\link{new_bids_class}} to
create new class definitions. All \code{bidsr} classes inherit this
abstract class, to provide consistent behavior.
}
