% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class010-bids_filename_registry.R
\name{parse_path_bids_entity}
\alias{parse_path_bids_entity}
\title{Parse 'BIDS' entities from file path}
\usage{
parse_path_bids_entity(path, auto_new = TRUE)
}
\arguments{
\item{path}{path to the entity file, recommended to input the absolute path
or relative path from the 'BIDS' root directory}

\item{auto_new}{whether to automatically generate class definition if
missing; default is true}
}
\value{
A \code{'bids_entity_file'} instance.
}
\description{
Parse 'BIDS' entities from file path
}
\examples{


path <- "anat/sub-001_t1w.nii.gz"

parsed_filename <- parse_path_bids_entity(path)
parsed_filename

parsed_filename$entities$sub$value

# get entity rules
parsed_filename$get_entity_rules()

# get class definition
definition <- S7::S7_class(parsed_filename)

# get entity rules from class
definition$get_entity_rules()


}
