% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa-generics-query_bids.R
\name{query_bids}
\alias{query_bids}
\title{Query 'BIDS'}
\usage{
query_bids(x, search_params, ...)
}
\arguments{
\item{x}{'BIDS' objects such as subject}

\item{search_params}{searching parameters, leave it blank to see help
documentations}

\item{...}{passed to down-stream methods}
}
\value{
A data table of query results
}
\description{
Query 'BIDS' project and analyze the files
}
