% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{set_props}
\alias{valid_eventually}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{S7}{\code{\link[S7:props]{set_props}}, \code{\link[S7:validate]{valid_eventually}}}
}}

