% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{download_bids_examples}
\alias{download_bids_examples}
\title{Download 'BIDS' example data-sets}
\usage{
download_bids_examples()
}
\value{
A local path to the example repository
}
\description{
See \url{https://github.com/bids-standard/bids-examples} for the full
repository.
}
\examples{

if(interactive()) {

  download_bids_examples()

}

}
