% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class020-bids_filename_registry.R
\name{new_bids_entity_file_class}
\alias{new_bids_entity_file_class}
\title{Class generator for 'BIDS' file class with entities}
\usage{
new_bids_entity_file_class(
  name,
  data_type,
  suffix,
  entity_rules = list(),
  description = character()
)
}
\arguments{
\item{name}{class name}

\item{data_type}{'BIDS' file data type}

\item{suffix}{file suffix}

\item{entity_rules}{named list, where each keys is a 'BIDS' entity key (see
\code{\link{bids_entity}}) and the corresponding value is a character vector
with length of 0, 1, or 2, containing the rules of such entity; see
Examples' below}

\item{description}{file class description}
}
\description{
Low-level function to generate file name definitions with entity
constraints; use \code{\link{parse_path_bids_entity}} instead.
The specification is at
\url{https://bids-specification.readthedocs.io/en/stable/common-principles.html#filenames}.
For bulk generating and registering 'BIDS' file classes, see
\code{\link{bids_entity_file_registry}}
}
\examples{


# see full table at BIDS specification
# en/stable/appendices/entity-table.html#behavioral-data
#
# generate class definition for "Behavioral Data"
# Entity: Subject Session Task Acquisition Run Recording
# Format:
#   sub-<label> ses-<label> task-<label>
#   acq-<label> run-<index> recording-<label>
# suffix: events
# requirement: REQUIRED OPTIONAL REQUIRED OPTIONAL OPTIONAL
#

behavior_event_file_def <- new_bids_entity_file_class(
  name = "bids_entity_file_beh_events",
  data_type = "beh",
  suffix = "events",
  entity_rules = list(
    sub = c("required", "label"),
    ses = c("optional", "label"),
    task = c("required", "label"),
    acq = c("optional", "label"),
    run = c("optional", "index"),
    recording = c("optional", "label")
  ),
  description = "Event timing information from a behavioral task."
)

file1 <- behavior_event_file_def(
  parent_directory = "sub-001/beh",
  sub = "001", task = "test", .extension = "tsv")

print(file1)

file.path("root/to/path", file1)

# get entity values
file1$get_bids_entity("task")

# parent directory
file1$parent_directory

file1$entities$run$value

# set entity values
file1$entities$run <- 2
file1$entities$blahblah <- "haha"

# also set format for index entities
file1$entities$run$index_format <- "\%03d"
file1


}
